+++ src/server/core/storage/document_internal.go
@@ -9,8 +9,11 @@ import (
 const MaxKeywordIndexSize = 1000
 
 var (
-	pendingWrites = map[string]*WorkspacePendingWrite{}
-	lastFlushTime = time.Now()
+	pendingWrites      = map[string]*WorkspacePendingWrite{}
+	lastFlushWriteTime = time.Now()
+
+	pendingDeletes      = map[string]*WorkspacePendingWrite{}
+	lastFlushDeleteTime = time.Now()
 )
 
 type RelatedDocs struct {
@@ -22,8 +25,7 @@ type WorkspacePendingWrite struct {
 	WorkspaceID string
 
 	// Map of keyword to document ids
-	Keywords  map[string]RelatedDocs
-	PathWords map[string]RelatedDocs
+	Keywords map[string]RelatedDocs
 }
 
 type flushPendingWritesTask struct {
@@ -33,6 +35,7 @@ type flushPendingWritesTask struct {
 func (t *flushPendingWritesTask) Run() {
 	// Flush pending writes to the database
 	flushPendingWrites(t.closing)
+	flushPendingDeletes(t.closing, MaxKeywordIndexSize)
 }
 
 func getPendingWrite(workspaceid string) *WorkspacePendingWrite {
@@ -41,7 +44,6 @@ func getPendingWrite(workspaceid string) *WorkspacePendingWrite {
 		wp = &WorkspacePendingWrite{
 			WorkspaceID: workspaceid,
 			Keywords:    make(map[string]RelatedDocs),
-			PathWords:   make(map[string]RelatedDocs),
 		}
 		pendingWrites[workspaceid] = wp
 	}
@@ -51,10 +53,10 @@ func getPendingWrite(workspaceid string) *WorkspacePendingWrite {
 
 // flushPendingWrites flushes the pending writes to the database
 func flushPendingWrites(closing bool) {
-	if !closing && time.Since(lastFlushTime) < 1*time.Second {
+	if !closing && time.Since(lastFlushWriteTime) < 1*time.Second {
 		return
 	}
-	lastFlushTime = time.Now()
+	lastFlushWriteTime = time.Now()
 
 	if closing {
 		log.Println("Final flushing pending writes...")
@@ -82,7 +84,7 @@ func flushPendingWrites(closing bool) {
 			delete(wp.Keywords, kw)
 
 			// delete empty workspace
-			if len(wp.Keywords) == 0 && len(wp.PathWords) == 0 {
+			if len(wp.Keywords) == 0 {
 				delete(pendingWrites, wp.WorkspaceID)
 			}
 		}
@@ -104,16 +106,70 @@ func updateKeywordIndexCached(workspaceid string, docid string, keywords []strin
 	}
 }
 
-func writeKeywordIndexCached(workspaceid string, keyword string, docids []string) {
-	cache := getPendingWrite(workspaceid)
-	cache.Keywords[keyword] = RelatedDocs{
-		DocIds:    append(cache.Keywords[keyword].DocIds, docids...),
-		UpdatedAt: time.Now(),
+// getPendingDelete returns the pending delete cache for the workspace
+// It will create a new cache if it does not exist
+func getPendingDelete(workspaceid string) *WorkspacePendingWrite {
+	wp := pendingDeletes[workspaceid]
+	if wp == nil {
+		wp = &WorkspacePendingWrite{
+			WorkspaceID: workspaceid,
+			Keywords:    make(map[string]RelatedDocs),
+		}
+		pendingDeletes[workspaceid] = wp
+	}
+
+	return wp
+}
+
+func flushPendingDeletes(closing bool, maxKeywordIndexSize int) {
+	if !closing && time.Since(lastFlushDeleteTime) < 1*time.Second {
+		return
+	}
+	lastFlushDeleteTime = time.Now()
+
+	if closing {
+		log.Println("Final flushing pending deletes...")
+		defer func() {
+			log.Println("Final fulshed pending deletes")
+		}()
+	}
+
+	batch := NewBatchWrite(db)
+
+	for _, wp := range pendingDeletes {
+		for kw, relatedDocs := range wp.Keywords {
+			// Skip the keyword if it has been updated in the last 2 seconds
+			// and has less than 50 documents
+			if !closing && len(relatedDocs.DocIds) < 50 && time.Since(relatedDocs.UpdatedAt) < 5*time.Second {
+				continue
+			}
+
+			removeDocumentsFromKeywordIndex(batch, wp.WorkspaceID, kw, relatedDocs.DocIds, maxKeywordIndexSize)
+			delete(wp.Keywords, kw)
+
+			// delete empty workspace
+			if len(wp.Keywords) == 0 {
+				delete(pendingDeletes, wp.WorkspaceID)
+			}
+		}
+	}
+
+	batch.Commit()
+}
+
+var removeKeywordsFromDocumentCached = func(workspaceid string, docid string, keywords []string) {
+	w := getPendingDelete(workspaceid)
+	for _, kw := range keywords {
+		// Add to delete cache to merge with other documents and flush later
+		w.Keywords[kw] = RelatedDocs{
+			DocIds:    append(w.Keywords[kw].DocIds, docid),
+			UpdatedAt: time.Now(),
+		}
 	}
 }
 
 // writeKeywordIndex writes a keyword to the database
-func writeKeywordIndex(batch BatchWrite, workspaceid string, kw string, docids []string, key []byte) {
+var writeKeywordIndex = func(batch BatchWrite, workspaceid string, kw string, docids []string, key []byte) {
 	content := EncodeKeywordIndexValue(docids)
 	if len(key) == 0 {
 		key = EncodeKeywordIndexKey(workspaceid, kw, len(docids))
@@ -121,21 +177,48 @@ func writeKeywordIndex(batch BatchWrite, workspaceid string, kw string, docids [
 	batch.Put(key, content)
 }
 
-// removeDocumentFromKeywordsIndex removes a document from the keywords index
+// removeDocumentsFromKeywordIndex removes a document from the keywords index
 // It will remove the document from the keywords index and rewrite the keyword with new docids
-func removeDocumentFromKeywordsIndex(batch BatchWrite, workspaceid string, kw string, removingDocid string) {
+func removeDocumentsFromKeywordIndex(batch BatchWrite, workspaceid string, kw string, removingDocids []string,
+	maxKeywordIndexSize int) {
 	if len(kw) == 0 {
 		log.Println("Warning: removing document from keywords index, but keyword is empty")
+		return
+	}
+
+	removings := map[string]struct{}{}
+	for _, id := range removingDocids {
+		if id != "" {
+			removings[id] = struct{}{}
+		}
+	}
+
+	if len(removings) == 0 {
+		log.Println("Warning: removing document from keywords index, but docid is empty")
 		return
 	}
 
 	keys := []string{}
 	docids := map[string]struct{}{}
 	db.Scan(EncodeKeywordIndexKeyPrefix(workspaceid, kw), func(key, value []byte) bool {
-		keys = append(keys, string(key))
+		changed := false
+		tmpids := []string{}
+
 		ids := DecodeKeywordIndexValue(string(value))
 		for _, id := range ids {
-			if id != removingDocid {
+			if _, ok := removings[id]; ok {
+				// remove the document from the keyword index
+				changed = true
+				continue
+			}
+			if id != "" {
+				tmpids = append(tmpids, id)
+			}
+		}
+
+		if changed || len(tmpids) < maxKeywordIndexSize/2 {
+			keys = append(keys, string(key))
+			for _, id := range tmpids {
 				docids[id] = struct{}{}
 			}
 		}
@@ -146,15 +229,21 @@ func removeDocumentFromKeywordsIndex(batch BatchWrite, workspaceid string, kw st
 	for len(docids) > 0 {
 		docs := []string{}
 		for id := range docids {
-			if len(docs) >= MaxKeywordIndexSize {
+			if len(docs) >= maxKeywordIndexSize {
 				break
 			}
 			docs = append(docs, id)
 			delete(docids, id)
 		}
 
+		var key string
+		if len(keys) > 0 {
+			key = keys[0]
+			keys = keys[1:]
+		}
+
 		// writeKeywordIndexCached(workspaceid, kw, docs)
-		writeKeywordIndex(batch, workspaceid, kw, docs, nil)
+		writeKeywordIndex(batch, workspaceid, kw, docs, []byte(key))
 		count++
 	}
 
@@ -271,10 +360,13 @@ func (t *updateDocumentsTask) Run() {
 			}
 		}
 
-		// Remove removed words from the keywords index
-		for _, kw := range removedWords {
-			removeDocumentFromKeywordsIndex(batch, t.WorkspaceID, kw, updatedDoc.ID)
-		}
+		removeKeywordsFromDocumentCached(t.WorkspaceID, updatedDoc.ID, removedWords)
+		/*
+			// Remove removed words from the keywords index
+			for _, kw := range removedWords {
+				removeDocumentFromKeywordsIndex(batch, t.WorkspaceID, kw, updatedDoc.ID)
+			}
+		*/
 
 		// Add new words to the keywords index
 		if len(newWords) > 0 {
@@ -324,12 +416,15 @@ func (t *deleteDocumentTask) Run() {
 		return
 	}
 
-	defer log.Println("Document '", doc.RelPath, "' deleted from workspace '", t.WorkspaceID, "'")
+	defer log.Printf("Document `%s` deleted from workspace `%s`", doc.RelPath, t.WorkspaceID)
 
-	// delete the document from the keywords
-	for _, kw := range doc.Words {
-		removeDocumentFromKeywordsIndex(batch, t.WorkspaceID, kw, t.DocId)
-	}
+	removeKeywordsFromDocumentCached(t.WorkspaceID, t.DocId, doc.Words)
+	/*
+		// delete the document from the keywords
+		for _, kw := range doc.Words {
+			removeDocumentFromKeywordsIndex(batch, t.WorkspaceID, kw, t.DocId)
+		}
+	*/
 
 	// delete the document meta and words
 	batch.Delete(EncodeDocumentMetaKey(t.WorkspaceID, t.DocId))